refactor: move pagination templates to common
