+++ src/factories/common.js
@@ -61,3 +61,18 @@ export function generateId(startingId, excludeList) {
 
   return startingId
 }
+
+export const pagination = {
+  number: 0,
+  size: 20,
+  totalPages: 1,
+  numberOfElements: 1,
+  totalElements: 1
+}
+export const pagination0 = {
+  number: 0,
+  size: 20,
+  totalPages: 1,
+  numberOfElements: 0,
+  totalElements: 0
+}
